backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.0
django-filter==23.5
Pillow==10.1.0
boto3==1.34.0
django-storages==1.14.2
python-decouple==3.8
django-extensions==3.2.3
gunicorn==21.2.0
whitenoise==6.6.0
django-compressor==4.4
django-imagekit==4.1.0
# AI Services
spacy==3.7.2
nltk==3.8.1
sentence-transformers==2.2.2
transformers==4.35.2
huggingface-hub==0.16.4,<0.17.0
scikit-learn==1.3.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.1.3
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0